[module] Verplaats ledenmemory naar een module en wat naar algemeen
