+++ lib/ledenmemory/view/LedenMemoryZijbalkView.php
@@ -1,0 +1,57 @@
+<?php
+/**
+ * LedenMemoryZijbalkView.php
+ *
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/04/2017
+ */
+class LedenMemoryZijbalkView implements View {
+
+	private $scores;
+	private $titel;
+
+	public function __construct($scores, $titel) {
+		$this->scores = $scores;
+		$this->titel = $titel;
+	}
+
+	function getTitel() {
+		return 'Topscores ' . $this->titel;
+	}
+
+	public function getBreadcrumbs() {
+		return null;
+	}
+
+	function getModel() {
+		return $this->scores;
+	}
+
+	function view() {
+		echo '<div id="zijbalk_ledenmemory_topscores"><div class="zijbalk-kopje"><a href="/forum/onderwerp/8017">';
+		echo $this->getTitel();
+		echo '</a></div>';
+		$first = true;
+		foreach ($this->getModel() as $score) {
+			echo '<div class="item">';
+			echo sprintf('%02d', floor($score->tijd / 60 % 60)); //minuten
+			echo ':';
+			echo sprintf('%02d', floor($score->tijd % 60)); //seconden
+			echo ' ';
+			if ($first) {
+				echo '<span class="cursief">';
+			}
+			echo ProfielModel::getLink($score->door_uid, 'civitas');
+			echo ' (';
+			echo $score->beurten;
+			echo ')';
+			if ($first) {
+				echo '</span>';
+			}
+			echo '</div>';
+			$first = false;
+		}
+		echo '</div>'; //einde wrapperdiv
+	}
+
+}
