--- lib/model/Paging.interface.php
@@ -1,22 +1,0 @@
-<?php
-
-/**
- * Paging.interface.php
- * 
- * @author P.W.G. Brussee <brussee@live.nl>
- *
- * Paginering van items.
- * 
- */
-interface Paging {
-
-	public function getAantalPerPagina();
-
-	public function setAantalPerPagina($aantal);
-
-	public function getHuidigePagina();
-
-	public function setHuidigePagina($pagina, $voor);
-
-	public function getAantalPaginas($voor);
-}
